cleaned up deployed code
