+++ cloudbuilds/cfgames.yaml
@@ -4,7 +4,7 @@ steps:
   - functions
   - deploy
   - cf_games_pipeline
-  - --source=.
+  - --source=./cf_games_pipeline
   - --trigger-topic=dailyscheduled
   - --runtime=python39
   - --memory=1024MB
