+++ cloudbuilds/cfgamestats.yaml
@@ -4,7 +4,7 @@ steps:
   - functions
   - deploy
   - cf_gamestats_pipeline
-  - --source=.
+  - --source=./cf_gamestats_pipeline
   - --trigger-topic=dailyscheduled
   - --runtime=python39
   - --memory=1024MB
