+++ cloudbuilds/uslocs.yaml
@@ -4,7 +4,7 @@ steps:
   - functions
   - deploy
   - uslocations_pipeline
-  - --source=.
+  - --source=./uslocations_pipeline
   - --trigger-topic=dailyscheduled
   - --runtime=python39
   - --memory=1024MB
