+++ cloudbuilds/weather.yaml
@@ -4,7 +4,7 @@ steps:
   - functions
   - deploy
   - weather_pipeline
-  - --source=.
+  - --source=./weather_pipeline
   - --trigger-topic=dailyscheduled
   - --runtime=python39
   - --memory=1024MB
